+++ prescia/components/intlControl.php
@@ -3,7 +3,7 @@
 
 class CintlControl {
 
-	var $parent = null;
+	public ?CPrescia $parent = null;
 	var $i18n = array(  #		dec  th  date	 date preg (opt 4 digit year)							   preg Y M D positions (0 all)
 						#						  12           3        45           6        7							 currency
 				'pt-br' => array(",",".","d/m/Y","(([0-9]{1,2})([^0-9]))(([0-9]{1,2})([^0-9]))([0-9]{2,4})",array(7,5,2),'BRL','R$'),
@@ -11,7 +11,7 @@ class CintlControl {
 				'es'    => array(",",".","m/d/Y","(([0-9]{1,2})([^0-9]))(([0-9]{1,2})([^0-9]))([0-9]{2,4})",array(7,2,5),'EUR','€'),
 				'jp'    => array(".",",","Y/m/d","(([0-9]{2,4})([^0-9]))(([0-9]{1,2})([^0-9]))([0-9]{1,2})",array(2,5,7),'JPY','¥'),
 				);
-	var $selectedCode = "";
+	public string $selectedCode = "";
 
 	function __construct(&$parent) {
 		$this->parent = &$parent;
@@ -200,4 +200,3 @@ class CintlControl {
 	} # langOut
 #-
 }
-